+++ app/Routes.hs
@@ -5,9 +5,12 @@ import Control.Monad (MonadPlus (mplus), msum)
 import Control.Monad.IO.Class (liftIO)
 import Controllers.Course as CoursesController (retrieveCourse, index, courseInfo, depts)
 import Controllers.Graph as GraphsController
+    ( graphResponse, index, getGraphJSON, graphImageResponse )
+import Controllers.Generate as GenerateController (generateResponse, findAndSavePrereqsResponse)
+import Controllers.Timetable as TimetableController
 import Data.Text.Lazy (Text)
 import Database.CourseInsertion (saveGraphJSON)
-import Database.CourseQueries (getGraphJSON, retrievePost)
+import Database.CourseQueries (retrievePost)
 import Happstack.Server
     ( serveDirectory,
       seeOther,
@@ -30,14 +33,8 @@ import Response
       privacyResponse,
       notFoundResponse,
       searchResponse,
-      generateResponse,
       postResponse,
-      loadingResponse,
-      gridResponse,
-      calendarResponse,
-      graphImageResponse,
-      exportTimetableImageResponse,
-      exportTimetablePDFResponse )
+      loadingResponse)
 
 routeResponses :: String -> Text -> Text -> ServerPartT IO Response
 routeResponses staticDir aboutContents privacyContents =
@@ -48,13 +45,13 @@ routeResponses staticDir aboutContents privacyContents =
 
 strictRoutes :: Text -> Text -> [ (String, ServerPart Response)] 
 strictRoutes aboutContents privacyContents = [
-    ("grid", gridResponse),
+    ("grid", TimetableController.gridResponse),
     ("graph", GraphsController.graphResponse),
     ("graph-generate", do method PUT
-                          GraphsController.findAndSavePrereqsResponse),
+                          GenerateController.findAndSavePrereqsResponse),
     ("image", look "JsonLocalStorageObj" >>= graphImageResponse),
-    ("timetable-image", lookText' "session" >>= \session -> look "courses" >>= exportTimetableImageResponse session),
-    ("timetable-pdf", look "courses" >>= \courses -> look "JsonLocalStorageObj" >>= exportTimetablePDFResponse courses),
+    ("timetable-image", lookText' "session" >>= \session -> look "courses" >>= TimetableController.exportTimetableImageResponse session),
+    ("timetable-pdf", look "courses" >>= \courses -> look "JsonLocalStorageObj" >>= TimetableController.exportTimetablePDFResponse courses),
     ("post", retrievePost),
     ("post-progress", postResponse),
     ("draw", drawResponse),
@@ -69,7 +66,7 @@ strictRoutes aboutContents privacyContents = [
     ("courses", CoursesController.index),
     ("course-info", lookText' "dept" >>= CoursesController.courseInfo),
     ("depts", CoursesController.depts),
-    ("calendar", look "courses" >>= calendarResponse),
+    ("calendar", look "courses" >>= TimetableController.calendarResponse),
     ("loading", lookText' "size" >>= loadingResponse),
     ("save-json", lookBS "jsonData" >>= \jsonStr -> lookText' "nameData" >>= \nameStr -> liftIO $ saveGraphJSON jsonStr nameStr)
     ]